(style) - search 페이지 비율 재조정
