+++ src/components/SearchPage/Slider.jsx
@@ -12,12 +12,16 @@ import Tip from "../../img/TipIcon.svg";
 
 import { ServiceData } from "./SliderContents";
 
-import Fadein from "../AboutPage/About_fadin"; //fadein 효과
+import Fadein from "../AboutPage/About_fadin";
 
 // Initialize required modules
 SwiperCore.use([FreeMode, Autoplay]);
 
 const StyledContainer = styled.div`
+  // 추가된 부분
+  margin-top: 580px;
+  // hard 코딩으로 간격 조정
+
   display: flex;
   flex-direction: column;
 
@@ -36,7 +40,7 @@ const TitleMain = styled.div`
   padding-right: 30px;
 
   font-family: EsaManru;
-  font-size: 40px;
+  font-size: 35px;
   font-weight: 500;
 `;
 const TitleSub = styled.div`
@@ -72,10 +76,12 @@ const CardTip = styled.div`
 `;
 
 const BigCont = styled.div`
-  padding: 150px 50px;
+  padding-top: 220px;
+  padding-left: 50px;
+  //padding: 350px 50px;
 
   font-family: EsaManru;
-  font-size: 140px;
+  font-size: 100px;
   font-weight: 700;
 `;
 
@@ -84,9 +90,13 @@ const Contents = styled.div`
   flex-direction: row-reverse;
   padding-right: 50px;
 
+  padding-top: 50px;
+
   font-family: Pretendard-Regular;
-  font-size: 20px;
+  font-size: 15px;
   font-weight: 500;
+
+  line-height: 26px;
 `;
 
 const UpBtn = styled.div`
@@ -94,7 +104,8 @@ const UpBtn = styled.div`
   flex-direction: column;
   align-items: center;
 
-  padding: 150px 0;
+  // padding: 150px 0;
+  padding: 50px;
 `;
 
 const ContainArrow = styled.div`
@@ -212,8 +223,8 @@ function Slider() {
 
       <Fadein>
         <BigCont>
-          Discover
-          <br /> New Journey
+          DISCOVER
+          <br /> NEW JOURNEY
         </BigCont>
       </Fadein>
