+++ src/style/Search_Banner.css
@@ -1,7 +1,3 @@
-/* div {
-  border: 1px solid blue;
-} */
-
 @font-face {
   font-family: "EsaManru";
   font-weight: 300;
@@ -74,76 +70,87 @@
 
 /*banner 전체 화면 구성*/
 .banner {
+  position: absolute;
+  top: 0;
   display: block;
+
+  width: 100%; /*풀스크린*/
+  /* width: 1920px; */
+  /* height: 5000px; */
 }
 
-/*About 소개 paige*/
-.about {
-  margin-top: 790px;
+.Enter {
+  /* background-color: gray; */
+  /* height: 1600px; */
+  height: 100%;
 
-  background-color: #ec4d27;
-
-  height: 400px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+  background-image: url(../img/Mainimage.png);
+  background-size: cover;
+  color: #ec4d27;
 }
 
-.AboutBox {
-  margin: 78px 453px;
-
-  width: 243px;
-  height: 179px;
-}
-
-.About1 {
-  font-family: "EsaManru";
-  color: white;
-
-  height: 70px;
-  font-size: 23px;
-  font-weight: 500;
-  line-height: 30px;
-  text-align: center;
-  letter-spacing: 2%;
-}
-
-.About2 {
+/*첫번째 글상자 양식*/
+.Enterphrase {
   font-family: "Pretendard-Regular";
-  color: white;
-  width: 250px;
-  height: 45px; /**/
 
+  margin-left: 50px;
+  padding-top: 120px;
+  height: 100px;
+
+  font-size: 42px;
   font-weight: 400;
-  font-size: 12px;
-  line-height: 14px;
-  text-align: center;
+  line-height: 50px;
 }
 
-.aboutBtn_container {
+/*Via-flight logo공간*/
+.Via_flight_logoNForm {
+  height: 250px;
+  margin-top: 67px;
+  margin-left: 50px;
+
+  .logo_text {
+    font-family: "EsaManru";
+
+    width: fit-content;
+    font-size: 135px; /*236px*/ /* #기준: [] */
+    font-weight: 700;
+    line-height: 143px;
+  }
+
+  .Flight_textNlogo {
+    display: flex;
+  }
+  .Flight_textNlogo > img {
+    height: 135px;
+    object-fit: scale-down;
+  }
+}
+
+.scrollIcon_container {
+  height: 260px;
+
   display: flex;
   justify-content: center;
 }
 
-.AboutBtn {
-  font-family: "Plaid";
-  color: #ec4d27;
+.scrollIcon_arrow {
+  margin-top: 150px;
+  /*img 가로, 세로 크기*/
 
-  font-size: 11px;
-
-  background-color: #f4f0e7;
-
-  border: 1px solid #ec4d27;
-  border-radius: 5px;
-
-  margin-top: 24px;
-  height: 42px;
-  width: 99px;
+  width: 20px;
+  height: 20px;
+  animation: motion 0.4s /*반복되는 시간*/ linear 0s /*효과 일정하게 반복*/
+    infinite alternate; /*애니매이션 효과 무한으로 반복; 상하로 움직이는 효과 */
 }
 
-.AboutBtn:hover {
-  color: rgba(244, 240, 231, 1);
-
-  background-color: #ec4d27;
-  border: 1px solid #f4f0e7;
+/*위-아래로 움직이는 효과*/
+@keyframes motion {
+  /*처음 위치*/
+  0% {
+    padding-top: 0px;
+  }
+  /*마지막 위치*/
+  100% {
+    padding-top: 10px;
+  }
 }
